# src/generated/openapi/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: My Project API
  version: 1.0.0
...
servers:
- description: Local Development Server
  url: https://api.platform.internal:8445
...
paths:
...
  /refresh:
    post:
      operationId: refresh
      responses:
        "204":
          description: Cookies refreshed
        "401":
          description: Invalid refresh token
      summary: Refresh authentication cookies
      tags:
      - auth
...